Fix getting the count of public repositories
Likely due to a GitHub bug, the API limits the `public_repos` count of a user to 50000. This commit replaces the usage of the `public_repos` property to querying the repositories endpoint and getting the total count from there. However, this requires extra API calls and is slower, so it should be reverted once the GitHub bug is fixed.

Consequently, this means that the current list of users by repositories is incorrect, but it will not be fixed for now.